rapatriate classes from BTDLinearSolver
