+++ SofaKernel/modules/Sofa.LinearAlgebra/CMakeLists.txt
@@ -8,6 +8,14 @@ set(HEADER_FILES
     ${SOFALINEARALGEBRASRC_ROOT}/init.h
     ${SOFALINEARALGEBRASRC_ROOT}/BaseMatrix.h
     ${SOFALINEARALGEBRASRC_ROOT}/BaseVector.h
+    ${SOFALINEARALGEBRASRC_ROOT}/BlockDiagonalMatrix.h
+    ${SOFALINEARALGEBRASRC_ROOT}/BlockDiagonalMatrix.inl
+    ${SOFALINEARALGEBRASRC_ROOT}/BlockVector.h
+    ${SOFALINEARALGEBRASRC_ROOT}/BlockVector.inl
+    ${SOFALINEARALGEBRASRC_ROOT}/BlocFullMatrix.h
+    ${SOFALINEARALGEBRASRC_ROOT}/BlocFullMatrix.inl
+    ${SOFALINEARALGEBRASRC_ROOT}/BTDMatrix.h
+    ${SOFALINEARALGEBRASRC_ROOT}/BTDMatrix.inl
     ${SOFALINEARALGEBRASRC_ROOT}/CompressedRowSparseMatrix.h
     ${SOFALINEARALGEBRASRC_ROOT}/DiagonalMatrix.h
     ${SOFALINEARALGEBRASRC_ROOT}/FullMatrix.h
