+++ SofaKernel/modules/Sofa.LinearAlgebra/src/sofa/linearalgebra/BlockVector.inl
@@ -19,22 +19,44 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
-#define SOFA_COMPONENT_LINEARSOLVER_BTDLINEARSOLVER_CPP
+#pragma once
+#include <sofa/linearalgebra/BlockVector.h>
 
-#include <SofaGeneralLinearSolver/BTDLinearSolver.inl>
-#include <sofa/core/ObjectFactory.h>
-#include <sofa/linearalgebra/BTDMatrix.inl>
-#include <sofa/linearalgebra/BlocFullMatrix.inl>
-#include <sofa/linearalgebra/BlockVector.inl>
-#include <SofaBaseLinearSolver/MatrixLinearSolver.inl>
-
-namespace sofa::component::linearsolver
+namespace sofa::linearalgebra
 {
 
-int BTDLinearSolverClass = core::RegisterObject("Linear system solver using Thomas Algorithm for Block Tridiagonal matrices")
-    .add< BTDLinearSolver<linearalgebra::BTDMatrix<6,double>, linearalgebra::BlockVector<6,double> > >(true)
-;
+template<std::size_t N, typename T>
+BlockVector<N, T>::BlockVector()
+{
+}
 
-template class SOFA_SOFAGENERALLINEARSOLVER_API BTDLinearSolver< linearalgebra::BTDMatrix<6, double>, linearalgebra::BlockVector<6, double> >;
+template<std::size_t N, typename T>
+BlockVector<N, T>::BlockVector(Index n)
+    : Inherit(n)
+{
+}
 
-} //namespace sofa::component::linearsolver
+template<std::size_t N, typename T>
+BlockVector<N, T>::~BlockVector()
+{
+}
+
+template<std::size_t N, typename T>
+typename BlockVector<N, T>::Bloc& BlockVector<N, T>::sub(Index i, Index)
+{
+    return (Bloc&)*(this->ptr()+i);
+}
+
+template<std::size_t N, typename T>
+const typename BlockVector<N, T>::Bloc& BlockVector<N, T>::asub(Index bi, Index) const
+{
+    return (const Bloc&)*(this->ptr()+bi*N);
+}
+
+template<std::size_t N, typename T>
+typename BlockVector<N, T>::Bloc& BlockVector<N, T>::asub(Index bi, Index)
+{
+    return (Bloc&)*(this->ptr()+bi*N);
+}
+
+} // namespace sofa::linearalgebra
