+++ modules/SofaGeneralLinearSolver/src/SofaGeneralLinearSolver/BTDLinearSolver.h
@@ -24,427 +24,14 @@
 
 #include <sofa/core/behavior/LinearSolver.h>
 #include <SofaBaseLinearSolver/MatrixLinearSolver.h>
-#include <SofaBaseLinearSolver/SparseMatrix.h>
+#include <sofa/linearalgebra/SparseMatrix.h>
+#include <sofa/linearalgebra/BTDMatrix.h>
+#include <sofa/linearalgebra/BlockVector.h>
 #include <cmath>
 #include <sofa/type/Mat.h>
 
 namespace sofa::component::linearsolver
 {
-
-/// Simple bloc full matrix container (used for InvMatrixType)
-template< std::size_t N, typename T>
-class BlocFullMatrix : public defaulttype::BaseMatrix
-{
-public:
-
-    enum { BSIZE = N };
-    typedef T Real;
-
-    class TransposedBloc{
-
-    public:
-        const type::Mat<BSIZE,BSIZE,Real>& m;
-
-        TransposedBloc(const sofa::type::Mat<BSIZE, BSIZE, Real>& m_a) : m(m_a){
-}
-
-        type::Vec<BSIZE,Real> operator*(const type::Vec<BSIZE,Real>& v)
-        {
-            return m.multTranspose(v);
-        }
-
-        type::Mat<BSIZE,BSIZE,Real> operator-() const
-        {
-            return -m.transposed();
-        }
-    };
-
-    class Bloc : public type::Mat<BSIZE,BSIZE,Real>
-    {
-    public:
-        Index Nrows() const;
-        Index Ncols() const;
-        void resize(Index, Index);
-        const T& element(Index i, Index j) const;
-        void set(Index i, Index j, const T& v);
-        void add(Index i, Index j, const T& v);
-        void operator=(const type::Mat<BSIZE,BSIZE,Real>& v)
-        {
-            type::Mat<BSIZE,BSIZE,Real>::operator=(v);
-        }
-        type::Mat<BSIZE,BSIZE,Real> operator-() const
-        {
-            return type::Mat<BSIZE,BSIZE,Real>::operator-();
-        }
-        type::Mat<BSIZE,BSIZE,Real> operator-(const type::Mat<BSIZE,BSIZE,Real>& m) const
-        {
-            return type::Mat<BSIZE,BSIZE,Real>::operator-(m);
-        }
-        type::Vec<BSIZE,Real> operator*(const type::Vec<BSIZE,Real>& v)
-        {
-            return type::Mat<BSIZE,BSIZE,Real>::operator*(v);
-        }
-        type::Mat<BSIZE,BSIZE,Real> operator*(const type::Mat<BSIZE,BSIZE,Real>& m)
-        {
-            return type::Mat<BSIZE,BSIZE,Real>::operator*(m);
-        }
-        type::Mat<BSIZE,BSIZE,Real> operator*(const Bloc& m)
-        {
-            return type::Mat<BSIZE,BSIZE,Real>::operator*(m);
-        }
-        type::Mat<BSIZE,BSIZE,Real> operator*(const TransposedBloc& mt)
-        {
-            return type::Mat<BSIZE,BSIZE,Real>::operator*(mt.m.transposed());
-        }
-        TransposedBloc t() const;
-        Bloc i() const;
-    };
-    typedef Bloc SubMatrixType;
-    // return the dimension of submatrices when requesting a given size
-    static Index getSubMatrixDim(Index);
-
-protected:
-    Bloc* data;
-    Index nTRow,nTCol;
-    Index nBRow,nBCol;
-    Index allocsize;
-
-public:
-
-    BlocFullMatrix();
-
-    BlocFullMatrix(Index nbRow, Index nbCol);
-
-    ~BlocFullMatrix() override;
-
-    Bloc* ptr() { return data; }
-    const Bloc* ptr() const { return data; }
-
-    const Bloc& bloc(Index bi, Index bj) const;
-
-    Bloc& bloc(Index bi, Index bj);
-
-    void resize(Index nbRow, Index nbCol) override;
-
-    Index rowSize(void) const override;
-
-    Index colSize(void) const override;
-
-    SReal element(Index i, Index j) const override;
-
-    const Bloc& asub(Index bi, Index bj, Index, Index) const;
-
-    const Bloc& sub(Index i, Index j, Index, Index) const;
-
-    Bloc& asub(Index bi, Index bj, Index, Index);
-
-    Bloc& sub(Index i, Index j, Index, Index);
-
-    template<class B>
-    void getSubMatrix(Index i, Index j, Index nrow, Index ncol, B& m);
-
-    template<class B>
-    void getAlignedSubMatrix(Index bi, Index bj, Index nrow, Index ncol, B& m);
-
-    template<class B>
-    void setSubMatrix(Index i, Index j, Index nrow, Index ncol, const B& m);
-
-    template<class B>
-    void setAlignedSubMatrix(Index bi, Index bj, Index nrow, Index ncol, const B& m);
-
-    void set(Index i, Index j, double v) override;
-
-    void add(Index i, Index j, double v) override;
-
-    void clear(Index i, Index j) override;
-
-    void clearRow(Index i) override;
-
-    void clearCol(Index j) override;
-
-    void clearRowCol(Index i) override;
-
-    void clear() override;
-
-    template<class Real2>
-    FullVector<Real2> operator*(const FullVector<Real2>& v) const
-    {
-        FullVector<Real2> res(rowSize());
-        for (Index bi=0; bi<nBRow; ++bi)
-        {
-            Index bj = 0;
-            for (Index i=0; i<BSIZE; ++i)
-            {
-                Real r = 0;
-                for (Index j=0; j<BSIZE; ++j)
-                {
-                    r += bloc(bi,bj)[i][j] * v[(bi + bj - 1)*BSIZE + j];
-                }
-                res[bi*BSIZE + i] = r;
-            }
-            for (++bj; bj<nBCol; ++bj)
-            {
-                for (Index i=0; i<BSIZE; ++i)
-                {
-                    Real r = 0;
-                    for (Index j=0; j<BSIZE; ++j)
-                    {
-                        r += bloc(bi,bj)[i][j] * v[(bi + bj - 1)*BSIZE + j];
-                    }
-                    res[bi*BSIZE + i] += r;
-                }
-            }
-        }
-        return res;
-    }
-
-
-    static const char* Name();
-};
-
-template< std::size_t N, typename T>
-class BlockVector : public FullVector<T>
-{
-public:
-    typedef FullVector<T> Inherit;
-    typedef T Real;
-    typedef typename Inherit::Index Index;
-
-    typedef typename Inherit::value_type value_type;
-    typedef typename Inherit::Size Size;
-    typedef typename Inherit::iterator iterator;
-    typedef typename Inherit::const_iterator const_iterator;
-
-    class Bloc : public type::Vec<N,T>
-    {
-    public:
-        Index Nrows() const { return N; }
-        void resize(Index) { this->clear(); }
-        void operator=(const type::Vec<N,T>& v)
-        {
-            type::Vec<N,T>::operator=(v);
-        }
-        void operator=(int v)
-        {
-            type::Vec<N,T>::fill((float)v);
-        }
-        void operator=(float v)
-        {
-            type::Vec<N,T>::fill(v);
-        }
-        void operator=(double v)
-        {
-            type::Vec<N,T>::fill(v);
-        }
-    };
-
-    typedef Bloc SubVectorType;
-
-public:
-
-    BlockVector();
-
-    explicit BlockVector(Index n);
-
-    virtual ~BlockVector();
-
-    const Bloc& sub(Index i, Index) const
-    {
-        return (const Bloc&)*(this->ptr()+i);
-    }
-
-    Bloc& sub(Index i, Index);
-
-    const Bloc& asub(Index bi, Index) const;
-
-    Bloc& asub(Index bi, Index);
-};
-
-/// Simple BTD matrix container
-template< std::size_t N, typename T>
-class BTDMatrix : public defaulttype::BaseMatrix
-{
-public:
-    enum { BSIZE = N };
-    typedef T Real;
-    typedef typename defaulttype::BaseMatrix::Index Index;
-
-    class TransposedBloc
-    {
-    public:
-        const type::Mat<BSIZE,BSIZE,Real>& m;
-        TransposedBloc(const type::Mat<BSIZE,BSIZE,Real>& m) : m(m) {}
-        type::Vec<BSIZE,Real> operator*(const type::Vec<BSIZE,Real>& v)
-        {
-            return m.multTranspose(v);
-        }
-        type::Mat<BSIZE,BSIZE,Real> operator-() const
-        {
-            type::Mat<BSIZE,BSIZE,Real> r;
-            for (Index i=0; i<BSIZE; i++)
-                for (Index j=0; j<BSIZE; j++)
-                    r[i][j]=-m[j][i];
-            return r;
-        }
-    };
-
-    class Bloc : public type::Mat<BSIZE,BSIZE,Real>
-    {
-    public:
-        Index Nrows() const { return BSIZE; }
-        Index Ncols() const { return BSIZE; }
-        void resize(Index, Index)
-        {
-            this->clear();
-        }
-        const T& element(Index i, Index j) const { return (*this)[i][j]; }
-        void set(Index i, Index j, const T& v) { (*this)[i][j] = v; }
-        void add(Index i, Index j, const T& v) { (*this)[i][j] += v; }
-        void operator=(const type::Mat<BSIZE,BSIZE,Real>& v)
-        {
-            type::Mat<BSIZE,BSIZE,Real>::operator=(v);
-        }
-        type::Mat<BSIZE,BSIZE,Real> operator-() const
-        {
-            type::Mat<BSIZE,BSIZE,Real> r;
-            for (Index i=0; i<BSIZE; i++)
-                for (Index j=0; j<BSIZE; j++)
-                    r[i][j]=-(*this)[i][j];
-            return r;
-        }
-        type::Mat<BSIZE,BSIZE,Real> operator-(const type::Mat<BSIZE,BSIZE,Real>& m) const
-        {
-            return type::Mat<BSIZE,BSIZE,Real>::operator-(m);
-        }
-        type::Vec<BSIZE,Real> operator*(const type::Vec<BSIZE,Real>& v)
-        {
-            return type::Mat<BSIZE,BSIZE,Real>::operator*(v);
-        }
-        type::Mat<BSIZE,BSIZE,Real> operator*(const type::Mat<BSIZE,BSIZE,Real>& m)
-        {
-            return type::Mat<BSIZE,BSIZE,Real>::operator*(m);
-        }
-        type::Mat<BSIZE,BSIZE,Real> operator*(const Bloc& m)
-        {
-            return type::Mat<BSIZE,BSIZE,Real>::operator*(m);
-        }
-        type::Mat<BSIZE,BSIZE,Real> operator*(const TransposedBloc& mt)
-        {
-            return type::Mat<BSIZE,BSIZE,Real>::operator*(mt.m.transposed());
-        }
-        TransposedBloc t() const
-        {
-            return TransposedBloc(*this);
-        }
-        Bloc i() const
-        {
-            Bloc r;
-            r.invert(*this);
-            return r;
-        }
-    };
-
-    typedef Bloc SubMatrixType;
-    typedef sofa::type::Mat<N,N,Real> BlocType;
-    typedef BlocFullMatrix<N,T> InvMatrixType;
-    // return the dimension of submatrices when requesting a given size
-    static Index getSubMatrixDim(Index) { return BSIZE; }
-
-protected:
-    Bloc* data;
-    Index nTRow,nTCol;
-    Index nBRow,nBCol;
-    Index allocsize;
-
-public:
-
-    BTDMatrix();
-
-    BTDMatrix(Index nbRow, Index nbCol);
-
-    ~BTDMatrix() override;
-
-    Bloc* ptr() { return data; }
-    const Bloc* ptr() const { return data; }
-
-    //Real* operator[](Index i)
-    //{
-    //    return data+i*pitch;
-    //}
-    const Bloc& bloc(Index bi, Index bj) const;
-
-    Bloc& bloc(Index bi, Index bj);
-
-    void resize(Index nbRow, Index nbCol) override;
-
-    Index rowSize(void) const override;
-
-    Index colSize(void) const override;
-
-    SReal element(Index i, Index j) const override;
-
-    const Bloc& asub(Index bi, Index bj, Index, Index) const;
-
-    const Bloc& sub(Index i, Index j, Index, Index) const;
-
-    Bloc& asub(Index bi, Index bj, Index, Index);
-
-    Bloc& sub(Index i, Index j, Index, Index);
-
-    template<class B>
-    void getSubMatrix(Index i, Index j, Index nrow, Index ncol, B& m);
-
-    template<class B>
-    void getAlignedSubMatrix(Index bi, Index bj, Index nrow, Index ncol, B& m);
-
-    template<class B>
-    void setSubMatrix(Index i, Index j, Index nrow, Index ncol, const B& m);
-
-    template<class B>
-    void setAlignedSubMatrix(Index bi, Index bj, Index nrow, Index ncol, const B& m);
-
-    void set(Index i, Index j, double v) override;
-
-    void add(Index i, Index j, double v) override;
-
-    void clear(Index i, Index j) override;
-
-    void clearRow(Index i) override;
-
-    void clearCol(Index j) override;
-
-    void clearRowCol(Index i) override;
-
-    void clear() override;
-
-    template<class Real2>
-    FullVector<Real2> operator*(const FullVector<Real2>& v) const
-    {
-        FullVector<Real2> res(rowSize());
-        for (Index bi=0; bi<nBRow; ++bi)
-        {
-            Index b0 = (bi > 0) ? 0 : 1;
-            Index b1 = ((bi < nBRow - 1) ? 3 : 2);
-            for (Index i=0; i<BSIZE; ++i)
-            {
-                Real r = 0;
-                for (Index bj = b0; bj < b1; ++bj)
-                {
-                    for (Index j=0; j<BSIZE; ++j)
-                    {
-                        r += data[bi*3+bj][i][j] * v[(bi + bj - 1)*BSIZE + j];
-                    }
-                }
-                res[bi*BSIZE + i] = r;
-            }
-        }
-        return res;
-    }
-
-    static const char* Name();
-};
-
-
 /// Linear system solver using Thomas Algorithm for Block Tridiagonal matrices
 ///
 /// References:
@@ -592,10 +179,7 @@ private:
 };
 
 #if  !defined(SOFA_COMPONENT_LINEARSOLVER_BTDLINEARSOLVER_CPP)
-extern template class SOFA_SOFAGENERALLINEARSOLVER_API BTDLinearSolver< BTDMatrix<6, double>, BlockVector<6, double> >;
-
-
-
+extern template class SOFA_SOFAGENERALLINEARSOLVER_API BTDLinearSolver< linearalgebra::BTDMatrix<6, double>, linearalgebra::BlockVector<6, double> >;
 #endif
 
 } //namespace sofa::component::linearsolver
